添加删除角色时的弹窗

+++ Unity/Codes/HotfixView/Demo/UI/DlgRoles/DlgRolesSystem.cs
@@ -94,6 +94,12 @@ namespace ET
 					return;
 				}
 
+				int result = await DlgTipsHelper.ShowTips(self.ZoneScene(),"确定要删除角色吗？");
+				if (result!=TipsResultType.Confirm)
+				{
+					return;
+				}
+
 				int errorCode = await LoginHelper.DeleteRole(self.ZoneScene());
 				if (errorCode!=ErrorCode.ERR_Success)
 				{
